third_party/WebKit/LayoutTests/imported/csswg-test/css-writing-modes-3/writing-mode-parsing-svg1-001-expected.txt

This is a testharness.js-based test.
PASS -webkit-writing-mode: lr 
PASS -webkit-writing-mode: lr-tb 
PASS -webkit-writing-mode: rl 
PASS -webkit-writing-mode: rl-tb 
FAIL -webkit-writing-mode: tb assert_equals: expected "vertical-rl" but got "horizontal-tb"
FAIL -webkit-writing-mode: tb-rl assert_equals: expected "vertical-rl" but got "horizontal-tb"
Harness: the test ran to completion.
